+++ src/Common/NetworkCore/Network/Pipeline/Outbound/DeliveryMethodOutboundPipelineStage.cs
@@ -64,7 +64,8 @@ public class DeliveryMethodOutboundPipelineStage : IOutboundPipelineStage
         { EventId.Core_Inventory_PickUp, DeliveryMethod.ReliableUnordered },
         { EventId.Core_Inventory_Drop, DeliveryMethod.ReliableUnordered },
         { EventId.Core_Inventory_DropAtPosition, DeliveryMethod.ReliableUnordered },
-        { EventId.Core_Inventory_Swap, DeliveryMethod.ReliableUnordered }
+        { EventId.Core_Inventory_Swap, DeliveryMethod.ReliableUnordered },
+        { EventId.Core_Inventory_UseItem, DeliveryMethod.ReliableUnordered }
     };
 
     public DeliveryMethodOutboundPipelineStage(ILogger<DeliveryMethodOutboundPipelineStage> logger)